+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog
 - Focus mode which hides the side bar and mades the main section of the page
 full width.
   - Option to enable and disable focus mode.
+  - Button to bottom of the page to toggle the focus mode.
+  - Option to enable and disabled the focus mode button.
 
 [v1.3.20] - 2020-06-25
 -----------------